backend/requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sympy==1.12
httpx==0.27.0
Pillow==10.3.0
python-dotenv==1.0.1
numpy==1.26.4
pydantic[email]==2.9.2
slowapi==0.1.9
redis==5.0.8
python-json-logger==2.0.7
